requirements.txt
# Core Django dependencies
Django==5.2.4
asgiref==3.8.1
sqlparse==0.5.3
tzdata==2025.2

# Image handling
Pillow==11.3.0

# Django forms and widgets
django-crispy-forms==2.3
crispy-bootstrap4==2024.1
django-widget-tweaks==1.5.0

# Production dependencies
gunicorn==23.0.0
whitenoise==6.8.2
dj-database-url==2.3.0
psycopg2-binary==2.9.12

# Environment variables
python-decouple==3.8

# Security
django-cors-headers==4.7.0